Add new event SET/UNSET, aul_open for auto-launch
1. SET/UNSET
SET will be used to set focus to a box.
Then the master will send mouse event to it directly.

UNSET will be used to unset a focus from the box.
Then the master will stop sending mouse events.

2. Auto launch
Previously, we use the aul_launch to launch an app via box.
So the app always got reset event when a user tries launch it.
This patch will do not send reset event to the app if it is already launched.
Instead of reset, this changes will send the resume event.

Issue: DCM-974

Change-Id: Ia1e515d0d7d6dd45144603f149c2b4586744b736