Use the vconf instead of timer for making connection to master.
If the master is not launched or ready to service,
Viewer should try to connect to it.

To try it, previously, we used timer.
So the viewer tries making connection after every N secs.

It can be harm to the system performance.
So this patch will change it to vconf.

So the master will change the vconf value,
Then the viewer notify its changing, and try to connect to it.

Change-Id: I6fa0a1467e223a917afbedf5aeb473ce9654bfd9